+++ plom/baseMessenger.py
@@ -127,7 +127,11 @@ class BaseMessenger:
         return self.session.patch(f"https://{self.server}" + url, *args, **kwargs)
 
     def start(self):
-        """Start the messenger session"""
+        """Start the messenger session.
+
+        Returns:
+            str: the version string of the server,
+        """
         if self.session:
             log.debug("already have an requests-session")
         else:
@@ -170,6 +174,22 @@ class BaseMessenger:
 
     def isStarted(self):
         return bool(self.session)
+
+    def get_server_version(self):
+        """The version info of the server.
+
+        Returns:
+            str: the version string of the server,
+
+        Exceptions:
+        """
+        with self.SRmutex:
+            try:
+                response = self.get("/Version")
+                response.raise_for_status()
+                return response.text
+            except requests.HTTPError as e:
+                raise PlomSeriousException(f"Some other sort of error {e}") from None
 
     # ------------------------
     # ------------------------
@@ -341,8 +361,7 @@ class BaseMessenger:
             this function converts them for us.
 
         Raises:
-            PlomServerNotReady: server does not yet have a version map,
-                e.g., b/c it has not been built, or server has no spec.
+            PlomAuthenticationException: login troubles.
         """
         with self.SRmutex:
             try:
@@ -354,8 +373,6 @@ class BaseMessenger:
             except requests.HTTPError as e:
                 if response.status_code == 401:
                     raise PlomAuthenticationException() from None
-                elif response.status_code in (404, 409):
-                    raise PlomServerNotReady(response.reason) from None
                 raise PlomSeriousException(f"Some other sort of error {e}") from None
         # JSON casts dict keys to str, force back to ints
         return undo_json_packing_of_version_map(response.json())
@@ -742,7 +759,7 @@ class BaseMessenger:
                 ) from None
             elif response.status_code == 410:
                 raise PlomBenignException(
-                    "That ID group of {} has not been scanned.".format(code)
+                    f"The ID page of {code} has not been scanned."
                 ) from None
             elif response.status_code == 409:
                 raise PlomSeriousException(