+++ plom/client/image_view_widget.py
@@ -13,7 +13,7 @@ from PyQt5.QtWidgets import (
     QGraphicsScene,
     QGraphicsView,
     QHBoxLayout,
-    QPushButton,
+    QToolButton,
     QVBoxLayout,
     QWidget,
 )
@@ -39,8 +39,10 @@ class ImageViewWidget(QWidget):
             these image files on disc or if the QPixmap will reads it once
             and stores it.  See Issue #1842.  For now, safest to assume
             you must maintain it.
-        has_reset_button (bool): whether to include a reset zoom button,
-            default: True.
+        has_controls (bool): include UI elements for zooming etc.
+            Default: True.
+        has_rotate_controls (bool): include UI elements for rotation.
+            Default: True.  Does nothing unless `has_controls` is True.
         compact (bool): whether to include a margin (default True) or
             not.  Correct choice will depend on parent but is probably
             only cosmetic.
@@ -53,54 +55,113 @@ class ImageViewWidget(QWidget):
         parent,
         image_data=None,
         *,
-        has_reset_button=True,
+        has_controls=True,
+        has_rotate_controls=True,
         compact=True,
         dark_background=False,
     ):
         super().__init__(parent)
         # Grab an examview widget (a interactive subclass of QGraphicsView)
-        self.view = ExamView(image_data, dark_background=dark_background)
+        self.view = _ExamView(image_data, dark_background=dark_background)
         self.view.setRenderHint(QPainter.Antialiasing, True)
         self.view.setRenderHint(QPainter.SmoothPixmapTransform, True)
         grid = QVBoxLayout()
         if compact:
             grid.setContentsMargins(0, 0, 0, 0)
         grid.addWidget(self.view, 1)
-        if has_reset_button:
-            resetB = QPushButton("&reset view")
+        grid.setSpacing(3)
+        self.zoomLockB = None
+        if has_controls:
+            resetB = QToolButton()
+            # resetB.setText("\N{Leftwards Arrow To Bar Over Rightwards Arrow To Bar}")
+            # resetB.setText("\N{Up Down Black Arrow}")
+            resetB.setText("reset")
+            resetB.setToolTip("reset zoom")
             resetB.clicked.connect(self.resetView)
-            # return won't click the button by default
-            resetB.setAutoDefault(False)
+            zoomInB = QToolButton()
+            zoomInB.setText("\N{Heavy Plus Sign}")
+            zoomInB.setToolTip("zoom in")
+            zoomInB.clicked.connect(self.zoomIn)
+            zoomOutB = QToolButton()
+            zoomOutB.setText("\N{Heavy Minus Sign}")
+            zoomOutB.setToolTip("zoom out")
+            zoomOutB.clicked.connect(self.zoomOut)
+            zoomLockB = QToolButton()
+            zoomLockB.setText(" \N{Lock} ")
+            zoomLockB.setCheckable(True)
+            zoomLockB.setChecked(False)
+            zoomLockB.clicked.connect(self.zoomLockToggle)
+            self.zoomLockB = zoomLockB
+            self._zoomLockUpdateTooltip()
             buttons = QHBoxLayout()
+            buttons.setContentsMargins(0, 0, 0, 0)
+            buttons.setSpacing(3)
+            buttons.addStretch(1)
             buttons.addWidget(resetB)
-            buttons.addSpacing(64)
+            buttons.addWidget(zoomInB)
+            buttons.addWidget(zoomOutB)
+            buttons.addWidget(zoomLockB)
+            if has_rotate_controls:
+                rotateB_cw = QToolButton()
+                rotateB_cw.setText("\N{Clockwise Open Circle Arrow}")
+                rotateB_cw.setToolTip("rotate clockwise")
+                rotateB_cw.clicked.connect(lambda: self.view.rotateImage(90))
+                rotateB_ccw = QToolButton()
+                rotateB_ccw.setText("\N{Anticlockwise Open Circle Arrow}")
+                rotateB_ccw.setToolTip("rotate counter-clockwise")
+                rotateB_ccw.clicked.connect(lambda: self.view.rotateImage(-90))
+                buttons.addSpacing(12)
+                buttons.addWidget(rotateB_cw)
+                buttons.addWidget(rotateB_ccw)
             buttons.addStretch(1)
             grid.addLayout(buttons)
+
         self.setLayout(grid)
-        # Store the current exam view as a qtransform
-        self.viewTrans = self.view.transform()
-        self.dx = self.view.horizontalScrollBar().value()
-        self.dy = self.view.verticalScrollBar().value()
 
     def updateImage(self, image_data):
         """Pass file(s) to the view to update the image"""
-        # first store the current view transform and scroll values
-        self.viewTrans = self.view.transform()
-        self.dx = self.view.horizontalScrollBar().value()
-        self.dy = self.view.verticalScrollBar().value()
         self.view.updateImages(image_data)
 
-        # re-set the view transform and scroll values
-        self.view.setTransform(self.viewTrans)
-        self.view.horizontalScrollBar().setValue(self.dx)
-        self.view.verticalScrollBar().setValue(self.dy)
+    def get_orientation(self):
+        """Report the sum of user-performed rotations."""
+        return self.view.theta
 
     def resizeEvent(self, whatev):
         """Seems to ensure image gets resize on window resize."""
+        if self.zoomLockB and self.zoomLockB.isChecked():
+            return
         self.view.resetView()
 
     def resetView(self):
+        if self.zoomLockB:
+            self.zoomLockB.setChecked(False)
+            self._zoomLockUpdateTooltip()
         self.view.resetView()
+
+    def zoomIn(self):
+        self.view.zoomIn()
+        self.zoomLockSetOn()
+
+    def zoomOut(self):
+        self.view.zoomOut()
+        self.zoomLockSetOn()
+
+    def zoomLockToggle(self):
+        if self.zoomLockB and not self.zoomLockB.isChecked():
+            # refit the view on untoggle
+            self.resetView()
+        self._zoomLockUpdateTooltip()
+
+    def zoomLockSetOn(self):
+        if self.zoomLockB:
+            self.zoomLockB.setChecked(True)
+            self._zoomLockUpdateTooltip()
+
+    def _zoomLockUpdateTooltip(self):
+        if self.zoomLockB.isChecked():
+            self.zoomLockB.setToolTip("zoom: locked")
+        else:
+            self.zoomLockB.setToolTip("zoom: fit on window resize")
 
     def forceRedrawOrSomeBullshit(self):
         """Horrid workaround when we cannot get proper redraws.
@@ -115,7 +176,7 @@ class ImageViewWidget(QWidget):
         QTimer.singleShot(32, self.view.resetView)
 
 
-class ExamView(QGraphicsView):
+class _ExamView(QGraphicsView):
     """Display images with some interaction: click-to-zoom/unzoom
 
     args:
@@ -143,6 +204,8 @@ class ExamView(QGraphicsView):
         self.scene = QGraphicsScene()
         self.imageGItem = QGraphicsItemGroup()
         self.scene.addItem(self.imageGItem)
+        # we track the total user-performed rotations in case caller is interested
+        self.theta = 0
         self.updateImages(image_data)
 
     def updateImages(self, image_data):
@@ -170,6 +233,10 @@ class ExamView(QGraphicsView):
             self.scene.removeItem(img)
         img = None
 
+        # we may use the viewing angle instead of rotating the item so reset
+        # if we have new images, even if they have non-zero orientation
+        self.resetTransform()
+
         if image_data is not None:
             x = 0
             for data in image_data:
@@ -189,6 +256,8 @@ class ExamView(QGraphicsView):
                 if pix.isNull():
                     raise ValueError(f"Could not read an image from {filename}")
                 rot = QTransform()
+                # if more than one image, its not well-defined which one theta gets
+                self.theta = data["orientation"]
                 rot.rotate(data["orientation"])
                 pix = pix.transformed(rot)
                 pixmap = QGraphicsPixmapItem(pix)
@@ -221,10 +290,18 @@ class ExamView(QGraphicsView):
         if (event.button() == Qt.RightButton) or (
             QGuiApplication.queryKeyboardModifiers() == Qt.ShiftModifier
         ):
-            self.scale(0.8, 0.8)
+            self.zoomOut()
         else:
-            self.scale(1.25, 1.25)
+            self.zoomIn()
         self.centerOn(event.pos())
+        # Unpleasant to grub in parent but want mouse events to lock zoom
+        self.parent().zoomLockSetOn()
+
+    def zoomOut(self):
+        self.scale(0.8, 0.8)
+
+    def zoomIn(self):
+        self.scale(1.25, 1.25)
 
     def resetView(self):
         """Reset the view to its reasonable initial state."""
@@ -232,4 +309,10 @@ class ExamView(QGraphicsView):
 
     def rotateImage(self, dTheta):
         self.rotate(dTheta)
-        self.resetView()
+        self.theta += dTheta
+        if self.theta == 360:
+            self.theta = 0
+        if self.theta == -90:
+            self.theta = 270
+        # Unpleasant to grub in parent but want to unlock zoom not just refit
+        self.parent().resetView()