+++ plom/create/__init__.py
@@ -24,6 +24,8 @@ from .buildDatabaseAndPapers import build_database, build_papers
 from .buildClasslist import process_classlist_file, get_demo_classlist
 from .upload_classlist import upload_classlist, upload_demo_classlist
 
+from .status import status
+
 from .push_pull_rubrics import upload_rubrics, download_rubrics
 from .push_pull_rubrics import upload_demo_rubrics
 from .push_pull_rubrics import upload_rubrics_from_file, download_rubrics_to_file
@@ -53,4 +55,5 @@ __all__ = [
     "version_map_from_file",
     "download_version_map",
     "save_version_map",
+    "status",
 ]