+++ plom/manager/selectrectangle.py
@@ -21,6 +21,7 @@ from PyQt5.QtWidgets import (
     QGraphicsScene,
     QGraphicsView,
     QGridLayout,
+    QLabel,
     QPushButton,
 )
 
@@ -38,7 +39,6 @@ class SelectRectangleWindow(QDialog):
         else:
             self.initUI([fnames])
         self.rectangle = None
-        self.whichFile = 0
         self.tool = "zoom"
 
     def initUI(self, fnames):
@@ -62,15 +62,27 @@ class SelectRectangleWindow(QDialog):
 
         self.resetB.setAutoDefault(False)  # return won't click the button by default.
 
-        # Layout simply
         grid = QGridLayout()
         grid.addWidget(self.view, 1, 1, 10, 6)
-        grid.addWidget(self.zoomB, 6, 20)
+        help = QLabel(
+            """
+            <p>Draw a box around the IDBox template.</p>
+            <ul>
+              <li>Bigger is better: a few cm above and below.</li>
+              <li>Only the top/bottom are currently used.</li>
+              <li>Tool won't work if you didn't use the IDBox template.</li>
+              <li>Tool does not fail gracefully: watch server log/stdout.</li>.
+            </ul>
+        """
+        )
+        help.setWordWrap(True)
+        grid.addWidget(help, 1, 19, 4, 3)
         grid.addWidget(self.rectB, 5, 20)
+        grid.addWidget(self.zoomB, 6, 20)
         grid.addWidget(self.delRectB, 7, 20)
         grid.addWidget(self.resetB, 20, 1)
         grid.addWidget(self.cancelB, 20, 20)
-        grid.addWidget(self.acceptB, 19, 20)
+        grid.addWidget(self.acceptB, 20, 21)
         self.setLayout(grid)
 
         self.rectB.animateClick()
@@ -82,6 +94,12 @@ class SelectRectangleWindow(QDialog):
             ErrorMessage("Error: no rectangle selected.").exec_()
             pass
         else:
+            t = self.view.imageGItem.boundingRect().top()
+            h = self.view.imageGItem.boundingRect().height() - t
+            self.top_bottom_values = (
+                max(0, (self.rectangle.top() - t) / h),
+                min(1, (self.rectangle.bottom() - t) / h),
+            )
             self.accept()
 
     def zoomTool(self):
@@ -147,6 +165,8 @@ class IDView(QGraphicsView):
                 # deal with jpeg exif rotations
                 qir.setAutoTransform(True)
                 pix = QPixmap(qir.read())
+                if pix.isNull():
+                    raise RuntimeError(f"Could not read an image from {fn}")
                 self.images[n] = QGraphicsPixmapItem(pix)
                 self.images[n].setTransformationMode(Qt.SmoothTransformation)
                 self.images[n].setPos(x, 0)
@@ -198,8 +218,6 @@ class IDView(QGraphicsView):
         if self.boxFlag:
             self.boxFlag = False
             self._parent.rectangle = self.boxItem.rect()
-            # legacy: left over from multiple id pages?
-            self._parent.whichFile = 0
             return
 
         """Left/right click to zoom in and out"""