+++ plom/managerMessenger.py
@@ -85,45 +85,41 @@ class ManagerMessenger(BaseMessenger):
     def appendTestToDB(self, test_number, vmap_for_test):
         """Ask server to append given test with the given versions.
 
-        TODO - fix up exceptions
-
         Returns:
             str: summary
 
         Raises:
-            PlomExistingDatabase: test already in database.
             PlomAuthenticationException: login troubles.
+            PlomServerNotReady: server has no spec.
+            PlomDatabaseCreationError: problems with version map or spec.
+            PlomConflict: row already exists or otherwise cannot create.
             PlomSeriousException: unexpected errors.
         """
-        self.SRmutex.acquire()
-        try:
-            response = self.put(
-                "/admin/appendTestToDB",
-                json={
-                    "user": self.user,
-                    "token": self.token,
-                    "test_number": test_number,
-                    "vmap_for_test": vmap_for_test,
-                },
-            )
-            response.raise_for_status()
-        except requests.HTTPError as e:
-            if response.status_code == 401:
-                raise PlomAuthenticationException() from None
-            if response.status_code == 403:
-                raise PlomAuthenticationException(response.reason) from None
-            if response.status_code == 406:
-                raise PlomDatabaseCreationError(response.reason) from None
-            if response.status_code == 409:
-                raise PlomExistingDatabase(response.reason) from None
-            if response.status_code == 400:
-                raise PlomServerNotReady(response.reason) from None
-            raise PlomSeriousException("Unexpected {}".format(e)) from None
-        finally:
-            self.SRmutex.release()
-
-        # JSON casts dict keys to str, force back to ints
-        return response.text
+        with self.SRmutex:
+            try:
+                response = self.put(
+                    "/admin/appendTestToDB",
+                    json={
+                        "user": self.user,
+                        "token": self.token,
+                        "test_number": test_number,
+                        "vmap_for_test": vmap_for_test,
+                    },
+                )
+                response.raise_for_status()
+                return response.text
+            except requests.HTTPError as e:
+                if response.status_code == 401:
+                    raise PlomAuthenticationException() from None
+                if response.status_code == 403:
+                    raise PlomAuthenticationException(response.reason) from None
+                if response.status_code == 406:
+                    raise PlomDatabaseCreationError(response.reason) from None
+                if response.status_code == 409:
+                    raise PlomConflict(response.reason) from None
+                if response.status_code == 400:
+                    raise PlomServerNotReady(response.reason) from None
+                raise PlomSeriousException("Unexpected {}".format(e)) from None
 
     def getGlobalPageVersionMap(self):
         with self.SRmutex:
@@ -1101,7 +1097,7 @@ class ManagerMessenger(BaseMessenger):
         finally:
             self.SRmutex.release()
 
-    def IDrunPredictions(self, rectangle, fileNumber, ignoreTimeStamp):
+    def IDrunPredictions(self, top, bottom, ignoreTimeStamp):
         self.SRmutex.acquire()
         try:
             response = self.post(
@@ -1109,8 +1105,8 @@ class ManagerMessenger(BaseMessenger):
                 json={
                     "user": self.user,
                     "token": self.token,
-                    "rectangle": rectangle,
-                    "fileNumber": fileNumber,
+                    "crop_top": top,
+                    "crop_bottom": bottom,
                     "ignoreStamp": ignoreTimeStamp,
                 },
             )
@@ -1123,6 +1119,8 @@ class ManagerMessenger(BaseMessenger):
         except requests.HTTPError as e:
             if response.status_code == 401:
                 raise PlomAuthenticationException() from None
+            if response.status_code == 403:
+                raise PlomAuthenticationException(response.reason) from None
             raise PlomSeriousException(f"Some other sort of error {e}") from None
         finally:
             self.SRmutex.release()