+++ plom/server/plomServer/routesUpload.py
@@ -841,23 +841,44 @@ class UploadHandler:
 
     @authenticate_by_token_required_fields(["user", "test_number", "vmap_for_test"])
     def appendTestToExamDatabase(self, data, request):
-        """Append given test to database using given version map."""
+        """Append given test to database using given version map.
+
+        Returns:
+            web.Response: 200 on success and a status message summarizing
+                the newly created row.
+                400 for server does not have spec.
+                401 for authentication, or 403 if not manager.
+                406 (unacceptable) for problems with version map or spec.
+                409 (conflict) for row already exists or otherwise cannot
+                be created.
+                500 for unexpected errors.
+        """
         if not data["user"] == "manager":
             raise web.HTTPForbidden(reason="Not manager")
+        spec = self.server.testSpec
+        if not spec:
+            raise web.HTTPBadRequest(reason="Server has no spec; cannot populate DB")
 
         # explicitly cast incoming vmap to ints
-        vmap = {int(q): int(v) for q, v in data["vmap_for_test"].items()}
+        try:
+            vmap = {int(q): int(v) for q, v in data["vmap_for_test"].items()}
+        except (TypeError, ValueError) as e:
+            raise web.HTTPNotAcceptable(
+                reason=f"Could not convert version map to int: {str(e)}"
+            ) from None
 
         try:
-            r, summary = self.server.appendTestToExamDatabase(data["test_number"], vmap)
-        except ValueError:
-            raise web.HTTPConflict(
-                reason="Attempt to build tests without contiguous numbers"
-            ) from None
-        if r:
-            return web.Response(text=summary, status=200)
-        else:
-            raise web.HTTPNotAcceptable(reason=summary)
+            summary = self.server.appendTestToExamDatabase(
+                spec, data["test_number"], vmap
+            )
+        except ValueError as e:
+            raise web.HTTPConflict(reason=str(e)) from None
+        except KeyError as e:
+            raise web.HTTPNotAcceptable(reason=str(e)) from None
+        except RuntimeError as e:
+            # uneasy about explicit 500, but these are unexpected
+            raise web.HTTPInternalServerError(reason=str(e)) from None
+        return web.Response(text=summary, status=200)
 
     @authenticate_by_token_required_fields([])
     def getGlobalPageVersionMap(self, data, request):
@@ -892,11 +913,11 @@ class UploadHandler:
 
         Returns:
             dict: keyed by question number.  Note keys will be strings b/c
-                of json limitations; you may need to convert back to int.
-                Fails with 409 if there is no version map.
+            of json limitations; you may need to convert back to int.
+            Fails with 409 if there is no such paper.
         """
         paper_idx = request.match_info["papernum"]
-        vers = self.server.getQuestionVersions(paper_idx)
+        vers = self.server.get_question_versions(paper_idx)
         if not vers:
             _msg = f"paper {paper_idx} does not (yet?) have a version map"
             log.warn(_msg)
@@ -909,24 +930,13 @@ class UploadHandler:
 
         Returns:
             dict: dict of dicts, keyed first by paper index then by
-                question number.  Both keys will become strings b/c of
-                json limitations; you may need to convert back to int.
-                Fails with 404/409 if there is no version map: 404 if
-                the server has no spec and 409 if the server has a spec
-                but the version map database has not been built yet.
+            question number.  Both keys will become strings b/c of
+            json limitations; you may need to convert back to int.
+            If the server does not yet have any database, the version
+            map will be empty.
         """
-        spec = self.server.testSpec
-        if not spec:
-            raise web.HTTPNotFound(reason="Server has no spec so no version map")
-        vers = {}
-        for paper_idx in range(1, spec["numberToProduce"] + 1):
-            ver = self.server.getQuestionVersions(paper_idx)
-            if not ver:
-                _msg = "There is no version map: have you built the database?"
-                log.warn(_msg)
-                raise web.HTTPConflict(reason=_msg)
-            vers[paper_idx] = ver
-        return web.json_response(vers, status=200)
+        vermap = self.server.get_all_question_versions()
+        return web.json_response(vermap, status=200)
 
     # Some more bundle things