+++ plom/server/plomServer/serverUpload.py
@@ -434,13 +434,17 @@ def initialiseExamDatabase(self, spec, vmap):
     return initialiseExamDatabaseFromSpec(spec, self.DB, vmap)
 
 
-def appendTestToExamDatabase(self, t, vmap_for_test):
-    return self.DB.addSingleTestToDB(self.testSpec, t, vmap_for_test)
+def appendTestToExamDatabase(self, spec, t, vmap_for_test):
+    return self.DB.addSingleTestToDB(spec, t, vmap_for_test)
 
 
 def getPageVersions(self, paper_idx):
     return self.DB.getPageVersions(paper_idx)
 
 
-def getQuestionVersions(self, paper_idx):
-    return self.DB.getQuestionVersions(paper_idx)
+def get_question_versions(self, paper_idx):
+    return self.DB.get_question_versions(paper_idx)
+
+
+def get_all_question_versions(self):
+    return self.DB.get_all_question_versions()